requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.30
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
sentence-transformers==2.7.0
chromadb==0.5.3
pypdf==4.2.0
httpx==0.27.0
streamlit==1.36.0
requests==2.32.3
pydantic[email]==2.7.4
python-dotenv==1.0.1
pytest==8.2.2
bcrypt==4.0.1
